requirements.txt
appdirs==1.4.4
arrow==0.15.6
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.2
black==20.8b1
blessed==1.17.8
boto3==1.14.12
botocore==1.17.12
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.34
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.4.0
django-filter==2.3.0
django-picklefield==3.0.1
django-q==1.2.4
django-ses==1.0.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.1
flake8==3.8.4
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==1.4.6
natsort==7.0.1
numpy==1.21.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.2.5
parso==0.7.1
pathspec==0.8.1
Pillow==9.2.0
pluggy==0.13.1
#psycopg2==2.9.4
psycopg2-binary==2.9.3
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pyotp==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.1
pytz==2020.1
redis==3.5.3
regex==2021.4.4
requests==2.24.0
rope==0.18.0
ruamel.yaml==0.16.10
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
stripe==2.56.0
toml==0.10.1
twilio==6.44.0
typed-ast==1.5.4
typing-extensions==3.7.4.3
ujson==4.0.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
wrapt==1.12.1
yapf==0.30.0
zope.event==4.5.0
zope.interface==5.2.0